requirements.txt
celery==5.3.4
Django==4.2.6
django-import-export==3.3.1
django-timezone-field==6.0.1
djangorestframework==3.14.0
gdown==4.7.1
pandas==2.1.1
redis==5.0.1